build: Add NMake makefiles for Visual Studio builds
This adds a set of NMake Makefiles that can be used to build pkg-config on
Windows using Visual Studio.  Note that, since the Visual Studio builds of
GLib does not use pkg-config, this NMake Makefile set does not support the
build of the bundled GLib sources, instead, having GLib built beforehand
is required.